+++ go_minimal_pool/shared/types.go
@@ -3,4 +3,4 @@ package shared
 type ParticipantId = uint32
 type EpochNumber = uint32
 type PoolId = uint32
-type PoolSize = uint8
+type PoolSize = uint32
